Move Sync Module, propagate MonadUnliftIO
