+++ unison-cli/src/Unison/Sync/HTTP.hs
@@ -4,12 +4,17 @@ import Control.Monad.Reader
 import qualified Network.HTTP.Client as HTTP
 import Unison.Sync
 import Unison.Sync.Types
+import Servant.Client
 
-newtype Sync a = Sync (ReaderT HTTP.Manager IO a)
+newtype SyncHTTP a = Sync (ReaderT HTTP.Manager IO a)
   deriving newtype (Functor, Applicative, Monad)
 
-instance MonadSync Sync where
+instance MonadSync SyncHTTP where
   getCausalHashForPath = _
   pushForce = _
   uploadToRepo = _
   downloadEntities = _
+
+runSyncHTTP :: MonadIO m => Auth.AuthorizedHTTPClient -> SyncHTTP a -> m a
+runSyncHTTP (AuthorizedHTTPClient manager) (Sync m) = liftIO $ do
+  runReaderT m manager
