+++ unison-core/unison-core1.cabal
@@ -57,7 +57,6 @@ library
       Unison.Util.Components
       Unison.Util.List
       Unison.Util.Monoid
-      Unison.Util.Tritraversable
       Unison.Var
       Unison.Var.RefNamed
       Unison.WatchKind
